gen-videos-vp8: use avconv instead of ffmpeg, and simply script
We now use avconv (from the libav project) instead of ffmpeg, since
libav is a more active fork of ffmpeg.

In addition to that, we no longer separate the audio from the video,
since we are capable to directly generate a video+audio for the
introduction, with 5 seconds of video showing the cover picture and 5
seconds of silence as audio. This allows to simplify the script.

We also mention which libav version has been used. We now do our
renderings inside a chroot to guarantee that all tools will remain at
the same version.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>