gen-html: width and height for LD video were inverted
The width and height variables for the low-res videos were inverted,
leading the generated HTML to create a link:

  450x800

for each low resolution video, making people believe that it is some
kind of portrait mode video.

So, invert the two variable values, and use the height to generate the
low-res video filename, since we name them 450p.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>